requirements.txt
anthropic
python-dotenv
pandas
scikit-learn
rdkit
